Modifiche da fare ancora
- Aggiungere la possibilità di vedere il nome nella navbar e gestire la grafica e intanto allinearsi con il progetto su digipad

+++ Sito_convention/Abbandona.php
@@ -1,16 +1,16 @@
 <?php
     include "Database.php";
     include "utilitis.php";
-    Controllo_Cookie();
+    Controllo_Cookie(false);
     session_start();
-    if((Controllo_Utente()) || !(isset($_POST['IdPart']))){
+    if((Controllo_Utente()) && (isset($_POST['IdPart']))){
         #l'utente ha cliccato ed è autenticato quindi aggiorno tempo sessione
         RefreshTempo();
         if (Database::connect()){
             $queryTab= "DELETE FROM Sceglie WHERE Sceglie.IDProgramma_fk = ? AND Sceglie.IDPart_fk = ?";
             $parametri=["ii",$_POST['QualeSpeech'],$_POST['IdPart']];
             $risultato=Database::executeQuery($queryTab,$parametri,false);
-            header("Location: ./home.php");
+            header("Location: ./index.php");
         }
     }
     else{
